+++ src/scriptdev2/scripts/eastern_kingdoms/sunwell_plateau/boss_kiljaeden.cpp
@@ -502,7 +502,7 @@ struct boss_kiljaedenAI : public Scripted_NoMovementAI, private DialogueHelper
 
     void GetAIInformation(ChatHandler& reader) override
     {
-        reader.PSendSysMessage("Kil'jaeden is currently in phase %u", m_uiPhase);
+        reader.PSendSysMessage("Kil'jaeden is currently in phase %u", uint32(m_uiPhase));
     }
 
     void JustDidDialogueStep(int32 iEntry) override
